+++ src/dir.c
@@ -821,11 +821,11 @@ print_file_list(FILE *out, const parray *files, const char *root)
 		if (root && strstr(path, root) == path)
 			path = GetRelativePath(path, root);
 
-		fprintf(out, "{\"path\":\"%s\", \"size\":\"%lu\",\"mode\":\"%u\","
+		fprintf(out, "{\"path\":\"%s\", \"size\":\"%d\",\"mode\":\"%u\","
 					 "\"is_datafile\":\"%u\", \"is_cfs\":\"%u\", \"crc\":\"%u\","
 					 "\"compress_alg\":\"%s\"",
-				path, (unsigned long) file->write_size, file->mode,
-				file->is_datafile?1:0, file->is_cfs?1:0, file->crc,
+				path, file->write_size, file->mode,
+				file->is_datafile ? 1 : 0, file->is_cfs ? 1 : 0, file->crc,
 				deparse_compress_alg(file->compress_alg));
 
 		if (file->is_datafile)
@@ -1032,7 +1032,7 @@ dir_read_file_list(const char *root, const char *file_txt)
 
 		file = pgFileInit(filepath);
 
-		file->write_size = (size_t) write_size;
+		file->write_size = (int) write_size;
 		file->mode = (mode_t) mode;
 		file->is_datafile = is_datafile ? true : false;
 		file->is_cfs = is_cfs ? true : false;
